pyproject.toml
[tool.poetry]
name = "meine-stadt-transparent"
version = "0.2.11"
description = "A website to bring municipal politics to citizens"
authors = ["Tobias Hößl <tobias@hoessl.eu>", "Konstantin Schütze <konstin@mailbox.org>"]
license = "MIT"
readme = "Readme.md"
include = ["mainapp/assets/bundles/**/*", "mainapp/migrations/*", "manage.py", "node_modules/pdfjs-dist/viewer/**/*"]
exclude = ["mainapp/assets/*"]
packages = [
    { include = "cms" },
    { include = "importer" },
    { include = "mainapp" },
    { include = "meine_stadt_transparent" },
]

[tool.poetry.dependencies]
Django = ">=3.0.8,<3.2.0"
PyPDF2 = "^1.26"
Wand = "^0.6.0"
cattrs = { version = "^1.0.0", optional = true }
django-allauth = ">=0.43,<0.45"
django-anymail = { version = "^8.1", extras = ["mailjet", "sendgrid"] }
django-decorator-include = ">=2.1,<4.0"
django-elasticsearch-dsl = ">=7.1,<7.3"
django-environ = "^0.4"
django-geojson = "^3.0"
django-settings-export = "^1.2"
django-simple-history = "^3.0"
django-webpack-loader = ">=0.6,<0.8"
django-widget-tweaks = "^1.4"
django_csp = "^3.4"
elasticsearch-dsl = "^7.3"
geoextract = "^0.3.1"
geopy = "^2.0.0"
gunicorn = "^20.0"
html2text = ">=2019.8,<2021.0"
icalendar = "^4.0"
jsonfield = "^3.1"
minio = "^5.0"
mysqlclient = ">=1.3,<3.0"
osm2geojson = "^0.1.28"
pgpy = { version = "^0.5.2", optional = true }
python = "^3.8"
python-dateutil = "^2.7"
python-slugify = "^5.0"
requests = "^2.22"
sentry-sdk = "^1.0.0"
splinter = "^0.14.0"
tqdm = "^4.29"
wagtail = ">=2.12,<2.13"

[tool.poetry.dev-dependencies]
black = "21.4b2"
django-debug-toolbar = "^3.0"
pytest-django = "^4.1"
responses = "^0.13.2"
selenium = "3.141.0"
tox = "^3.20"

[tool.poetry.extras]
pgp = ["pgpy"]
import-json = ["cattrs"]

[tool.poetry.scripts]
mst-manage = 'manage:main'

[tool.black]
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.venv
  | \.tox
  | mainapp/migrations
  | cms/migrations
  | importer/migrations
  | node_modules
  | static
)/
'''

[build-system]
requires = ["poetry~=1.1.0"]
build-backend = "poetry.masonry.api"
